chore: update icon tools to typescript
