+++ packages/paste-icons/.prettierrc
@@ -6,5 +6,13 @@
   "bracketSpacing": false,
   "singleQuote": true,
   "jsxBracketSameLine": false,
-  "tabWidth": 2
+  "tabWidth": 2,
+  "overrides": [
+    {
+      "files": "*.json",
+      "options": {
+        "parser": "json"
+      }
+    }
+  ]
 }
