+++ packages/paste-icons/tools/constants.ts
@@ -1,0 +1,6 @@
+import {join} from 'path';
+
+export const SVG_PATH = join(__dirname, '../svg');
+export const REACT_PATH = join(__dirname, '../src');
+
+export const BLOCKLIST_FILES = ['.DS_Store', 'helpers', 'types'];
