+++ packages/paste-icons/tools/templates/jsonTemplate.ts
@@ -1,14 +1,10 @@
-const jsonTemplate = iconList => {
+export const jsonTemplate = (iconList: string[] | undefined): string => {
   return JSON.stringify(
-    iconList.map(iconName => ({
+    iconList?.map((iconName: string) => ({
       name: iconName,
       svg: `/svg/${iconName}.svg`,
       cjs: `/cjs/${iconName}.js`,
       esm: `/esm/${iconName}.js`,
     }))
   );
-};
-
-module.exports = {
-  jsonTemplate,
 };
